+++ node/package.json
@@ -9,7 +9,7 @@
     "@types/jest": "^27.4.1",
     "@types/node": "^12.0.0",
     "@types/ramda": "types/npm-ramda#dist",
-    "@vtex/api": "6.45.10",
+    "@vtex/api": "6.45.11-beta",
     "@vtex/test-tools": "^3.4.3",
     "@vtex/tsconfig": "^0.6.0",
     "prettier": "^2.5.1",
