+++ node/service.json
@@ -25,16 +25,15 @@
     "mappMessages": {
       "path": "/_v/app/vtex-mapp-cloud/mappMessages",
       "public": true
-    },
-    "getProducts": {
-      "path": "/_v/app/vtex-mapp-cloud/getProducts",
-      "public": true
     }
   },
   "events": {
     "orderStatusOnChange": {
       "sender": "vtex.orders-broadcast",
       "topics": ["on-order-completed", "cancel"]
+    },
+    "skuChange": {
+      "keys": ["broadcaster.notification"]
     }
   }
 }
