+++ node/src/clients/pricing.ts
@@ -1,7 +1,7 @@
 import type {InstanceOptions, IOContext} from "@vtex/api"
 import {ExternalClient} from "@vtex/api"
 
-import type {PriceData} from "../typings/product"
+import type {VtexPriceData} from "../typings/vtex"
 
 export default class PricingClient extends ExternalClient {
   constructor(context: IOContext, options?: InstanceOptions) {
@@ -18,7 +18,7 @@ export default class PricingClient extends ExternalClient {
     })
   }
 
-  public getPrice(skuId: string | number): Promise<PriceData> {
-    return this.http.get<PriceData>(`/prices/${skuId}`)
+  public getPrice(skuId: string | number): Promise<VtexPriceData> {
+    return this.http.get<VtexPriceData>(`/prices/${skuId}`)
   }
 }
