--- node/src/typings/mapp-connect-catalog.d.ts
@@ -1,12 +1,0 @@
-export interface MappConnectCatalogItem {
-  productSKU: string
-  productName: string
-  productPrice: number
-  stockTotal: number
-  productURL: string
-  imageURL: string
-  zoomImageURL: string
-  brand: string
-  category: string
-  description: string
-}
