--- node/src/typings/order.d.ts
@@ -1,54 +1,0 @@
-interface Order {
-  orderId: string
-  clientProfileData: ClientProfileData
-  creationDate: string
-  value: number
-  totals: Total[]
-  items: VtexOrderItem[]
-  status: string
-  statusDescription: string
-  sequence: string
-  salesChannel: string
-  affiliateId: string | null
-  origin: string | null
-  authorizedDate: string | null
-  invoicedDate: string | null
-  marketingData?: MarketingData | null
-  shippingData: ShippingData
-  checkedInPickupPointId: number
-  storePreferencesData: StorePreferencesData
-}
-interface StorePreferencesData {
-  currencyCode: string
-}
-interface ClientProfileData {
-  email: string
-  firstName: string
-  lastName: string
-  documentType: string
-  document: string | null
-  phone: string
-  corporateName: string | null
-  tradeName: string | null
-  corporateDocument: string | null
-  stateInscription: string | null
-  corporatePhone: string | null
-  isCorporate: boolean
-  userProfileId: string | null
-}
-interface MarketingData {
-  coupon: string
-}
-interface Total {
-  id: string
-  name: string
-  value: string
-}
-interface VtexOrderItem {
-  quantity: number
-  id: string
-  productId: string | null
-  price: number
-  name: string
-  sellerSku: string
-}
