--- node/src/typings/user.d.ts
@@ -1,19 +1,0 @@
-interface User {
-  id: string
-  userId: string
-  email: string
-  isNewsletterOptIn: boolean
-  profilePicture?: string
-  isCorporate: boolean
-  homePhone?: string
-  phone?: string
-  businessPhone?: string
-  stateRegistration?: string
-  firstName?: string
-  lastName?: string
-  localeDefault?: string
-  birthDate?: string
-  corporateName?: string
-  gender?: string
-  birthDateMonth?: string
-}
